Move to .patch.json system
