+++ src/main/resources/patches/entity/arrow.entity.patch.json
@@ -1,0 +1,9 @@
+{
+  "minecraft:client_entity": {
+    "description": {
+      "textures": {
+        "spectral": "textures/geyser/entity/arrow/spectral_arrow"
+      }
+    }
+  }
+}
