+++ src/main/resources/patches/entity/evocation_illager.entity.patch.json
@@ -1,0 +1,1 @@
+{"minecraft:client_entity": {"description": {"textures": {"illusioner": "textures/geyser/entity/illager/illusioner"}}}}
