+++ src/main/resources/patches/entity/rabbit.entity.patch.json
@@ -1,0 +1,1 @@
+{"minecraft:client_entity": {"description": {"textures": {"caerbannog": "textures/geyser/entity/rabbit/caerbannog"}}}}
