+++ src/main/resources/patches/particles/basic_bubble_manual.patch.json
@@ -1,0 +1,1 @@
+{"particle_effect": {"components": {"minecraft:particle_expire_if_not_in_blocks": []}}}
