+++ src/main/resources/patches/particles/trial_spawner_detection.particle.patch.json
@@ -1,0 +1,1 @@
+{"particle_effect": {"components": {"minecraft:emitter_rate_instant": {"num_particles": "1"}, "minecraft:emitter_shape_box": {"offset": [0, 0, 0]}}}}
