+++ src/main/resources/patches/render_controllers/arrow.render_controllers.patch.json
@@ -1,0 +1,1 @@
+{"render_controllers": {"controller.render.arrow": {"textures": ["q.is_bribed ? texture.spectral : texture.default"]}}}
